* numeric.c (flo_eq): supress a warning on VC++ for x64.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@26050 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ numeric.c
@@ -904,7 +904,7 @@ flo_eq(VALUE x, VALUE y)
 
     switch (TYPE(y)) {
       case T_FIXNUM:
-	b = FIX2LONG(y);
+	b = (double)FIX2LONG(y);
 	break;
       case T_BIGNUM:
 	b = rb_big2dbl(y);
